backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
bcrypt==5.0.0
boto3==1.40.47
botocore==1.40.47
certifi==2025.10.5
charset-normalizer==3.4.3
chonkie==1.3.1
click==8.1.8
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.118.0
frozenlist==1.8.0
googlemaps==4.10.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
jmespath==1.0.1
multidict==6.7.0
numpy==2.0.2
passlib==1.7.4
pgvector==0.4.1
propcache==0.4.0
psycopg==3.2.10
psycopg-binary==3.2.10
pyasn1==0.6.1
pydantic==2.12.0
pydantic-core==2.41.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
starlette==0.48.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==1.26.20
uvicorn==0.37.0
yarl==1.22.0
